Added V49 hash, removed MoreCompany dependency
MoreCompany version 1.7.3 introduced a visual glitch that causes the left eye to glitch in VR while on the main game screen.

Unfortunately the author of MoreCompany has not pushed any updates to GitHub so the code cannot be reviewed. This patch removes the MoreCompany dependency to allow LCVR to be playable (with some functionality removed) in V49.

MoreCompany can be installed separately to allow more than 4 players to join a game with LCVR.